customized webpack config to accept jsx
