+++ src/App.jsx
@@ -1,6 +1,7 @@
 // import validator from 'validator';
 import React from 'react';
 import ReactDOM from 'react-dom';
+import AddOption from './components/AddOption';
 
 class IndecisionApp extends React.Component {
   constructor(props) {
@@ -174,42 +175,7 @@ const Options = (props) => {
 };
 
 
-class AddOption extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      error: undefined,
-    }
-    this.localHandleAddOption = this.localHandleAddOption.bind(this)
-  }
 
-  localHandleAddOption (e) {
-    e.preventDefault();
-
-    const option = e.target.elements.option.value.trim();
-    const error = this.props.handleAddOption(option);
-
-    this.setState(() => ({
-        error: error,
-    }));
-
-    if (!error) {
-      e.target.elements.option.value = '';
-    }
-  }
-
-  render() {
-    return (
-      <div>
-        <form onSubmit={this.localHandleAddOption}>
-          <input type="text" name="option" />
-          <button>Add Option</button>
-        </form>
-        {this.state.error && <p>{this.state.error}</p>}
-      </div>
-    );
-  }
-}
 
 const Option = (props) => {
   // console.log('Option props: ', props)
