+++ src/components/AddOption.jsx
@@ -1,0 +1,38 @@
+import React from 'react';
+
+export default class AddOption extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      error: undefined,
+    }
+    this.localHandleAddOption = this.localHandleAddOption.bind(this)
+  }
+
+  localHandleAddOption (e) {
+    e.preventDefault();
+
+    const option = e.target.elements.option.value.trim();
+    const error = this.props.handleAddOption(option);
+
+    this.setState(() => ({
+        error: error,
+    }));
+
+    if (!error) {
+      e.target.elements.option.value = '';
+    }
+  }
+
+  render() {
+    return (
+      <div>
+        <form onSubmit={this.localHandleAddOption}>
+          <input type="text" name="option" />
+          <button>Add Option</button>
+        </form>
+        {this.state.error && <p>{this.state.error}</p>}
+      </div>
+    );
+  }
+}
