--- src/utils.js
@@ -1,9 +1,0 @@
-console.log('utils is running');
-
-const square = (x) => x * x;
-
-const add = (a, b) => a + b;
-
-const subtract = (a, b) => a - b;
-
-export {square, add, subtract as default}
